+++ src/commonMain/kotlin/space/kscience/gdml/Gdml.kt
@@ -3,13 +3,12 @@
 
 package space.kscience.gdml
 
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.Transient
-import kotlinx.serialization.UseSerializers
+import kotlinx.serialization.*
 import nl.adaptivity.xmlutil.serialization.UnknownChildHandler
 import nl.adaptivity.xmlutil.serialization.XML
 import nl.adaptivity.xmlutil.serialization.XmlSerialName
+import kotlin.reflect.KType
+import kotlin.reflect.typeOf
 
 @DslMarker
 public annotation class GdmlApi
@@ -72,17 +71,12 @@ public class Gdml {
     public inline fun <reified T : GdmlGroup> getGroup(ref: String): T? =
         containers.filterIsInstance<GdmlStructure>().mapNotNull { it.get<T>(ref) }.singleOrNull()
 
-    public var world: GdmlGroup
-        get() = setup.world?.resolve(this)
-            ?: error("The GDML structure does not contain a world volume")
+    public var world: GdmlRef<GdmlGroup>
+        get() = setup.world ?: error("The GDML structure does not contain a world volume")
         set(value) {
-            //Add world element if it is not registered
-            structure {
-                if (getItem(value.name) == null) {
-                    add(value)
-                }
-            }
-            setup.world = value.ref()
+            //Check that the world volume is registered
+            if (getGroup<GdmlGroup>(value.ref) == null) error("World element is not registered")
+            setup.world = value
         }
 
 
@@ -157,6 +151,26 @@ public sealed class GdmlContainer<T : GdmlNode> {
     }
 
     /**
+     * A counter used for auto-naming
+     */
+    @Transient
+    private var autoNameCounter = 0
+
+    @OptIn(ExperimentalSerializationApi::class)
+    @PublishedApi
+    internal fun autoName(type: KType): String {
+        val serialName = kotlinx.serialization.serializer(type).descriptor.serialName
+        return "$serialName[${autoNameCounter++}]"
+    }
+
+    /**
+     * Use [providedName] if it is not null or generate unique automatic name
+     */
+    @PublishedApi
+    internal inline fun <reified T : GdmlNode> resolveName(providedName: String?): String =
+        providedName ?: autoName(typeOf<T>())
+
+    /**
      * Get an Item from th container
      */
     public fun getItem(ref: String): T? = cache.getOrPut(ref) { content.find { it.name == ref } }
@@ -175,14 +189,14 @@ public class GdmlDefineContainer : GdmlContainer<GdmlDefine>() {
 
     @GdmlApi
     public inline fun position(
-        name: String,
         x: Number = 0f,
         y: Number = 0f,
         z: Number = 0f,
+        name: String? = null,
         block: GdmlPosition.() -> Unit = {},
     ): GdmlRef<GdmlPosition> {
         val position = GdmlPosition().apply(block).apply {
-            this.name = name
+            this.name = resolveName<GdmlPosition>(name)
             this.x = x
             this.y = y
             this.z = z
@@ -193,14 +207,14 @@ public class GdmlDefineContainer : GdmlContainer<GdmlDefine>() {
 
     @GdmlApi
     public inline fun rotation(
-        name: String,
         x: Number = 0f,
         y: Number = 0f,
         z: Number = 0f,
+        name: String? = null,
         block: GdmlRotation.() -> Unit = {},
     ): GdmlRef<GdmlRotation> {
         val rotation = GdmlRotation().apply(block).apply {
-            this.name = name
+            this.name = resolveName<GdmlRotation>(name)
             this.x = x
             this.y = y
             this.z = z
@@ -215,14 +229,14 @@ public class GdmlDefineContainer : GdmlContainer<GdmlDefine>() {
 public class GdmlMaterialContainer : GdmlContainer<GdmlMaterial>() {
     override val content: MutableList<GdmlMaterial> = ArrayList()
 
-    public inline fun isotope(name: String, build: GdmlIsotope.() -> Unit = {}): GdmlRef<GdmlIsotope> =
-        register(GdmlIsotope(name).apply(build))
+    public inline fun isotope(name: String? = null, build: GdmlIsotope.() -> Unit = {}): GdmlRef<GdmlIsotope> =
+        register(GdmlIsotope(resolveName<GdmlIsotope>(name)).apply(build))
 
-    public inline fun element(name: String, build: GdmlElement.() -> Unit = {}): GdmlRef<GdmlElement> =
-        register(GdmlElement(name).apply(build))
+    public inline fun element(name: String? = null, build: GdmlElement.() -> Unit = {}): GdmlRef<GdmlElement> =
+        register(GdmlElement(resolveName<GdmlIsotope>(name)).apply(build))
 
-    public inline fun composite(name: String, build: GdmlComposite.() -> Unit = {}): GdmlRef<GdmlComposite> =
-        register(GdmlComposite(name).apply(build))
+    public inline fun composite(name: String? = null, build: GdmlComposite.() -> Unit = {}): GdmlRef<GdmlComposite> =
+        register(GdmlComposite(resolveName<GdmlIsotope>(name)).apply(build))
 
     public companion object {
         public val defaultMaterial: GdmlElement = GdmlElement("default")
@@ -236,121 +250,126 @@ public class GdmlSolidContainer : GdmlContainer<GdmlSolid>() {
 
     @GdmlApi
     public inline fun box(
-        name: String,
         x: Number,
         y: Number,
         z: Number,
+        name: String? = null,
         block: GdmlBox.() -> Unit = {},
-    ): GdmlRef<GdmlBox> = register(GdmlBox(name, x, y, z).apply(block))
+    ): GdmlRef<GdmlBox> = register(GdmlBox(resolveName<GdmlBox>(name), x, y, z).apply(block))
 
     @GdmlApi
     public inline fun sphere(
-        name: String,
         rmax: Number,
+        name: String? = null,
         block: GdmlSphere.() -> Unit = {},
-    ): GdmlRef<GdmlSphere> = register(GdmlSphere(name, rmax = rmax).apply(block))
+    ): GdmlRef<GdmlSphere> = register(GdmlSphere(resolveName<GdmlSphere>(name), rmax = rmax).apply(block))
 
 
     @GdmlApi
     public inline fun orb(
-        name: String,
         r: Number,
+        name: String? = null,
         block: GdmlOrb.() -> Unit = {},
-    ): GdmlRef<GdmlOrb> = register(GdmlOrb(name, r).apply(block))
+    ): GdmlRef<GdmlOrb> = register(GdmlOrb(resolveName<GdmlOrb>(name), r).apply(block))
 
     @GdmlApi
     public inline fun ellipsoid(
-        name: String,
         ax: Number,
         by: Number,
         cz: Number,
+        name: String? = null,
         block: GdmlEllipsoid.() -> Unit = {},
-    ): GdmlRef<GdmlEllipsoid> = register(GdmlEllipsoid(name, ax, by, cz).apply(block))
+    ): GdmlRef<GdmlEllipsoid> = register(GdmlEllipsoid(resolveName<GdmlEllipsoid>(name), ax, by, cz).apply(block))
 
 
     @GdmlApi
     public inline fun eltube(
-        name: String,
         dx: Number,
         dy: Number,
         dz: Number,
+        name: String? = null,
         block: GdmlElTube.() -> Unit = {},
-    ): GdmlRef<GdmlElTube> = register(GdmlElTube(name, dx, dy, dz).apply(block))
+    ): GdmlRef<GdmlElTube> = register(GdmlElTube(resolveName<GdmlElTube>(name), dx, dy, dz).apply(block))
 
     @GdmlApi
     public inline fun elcone(
-        name: String,
         dx: Number,
         dy: Number,
         zmax: Number,
         zcut: Number,
+        name: String? = null,
         block: GdmlElCone.() -> Unit = {},
-    ): GdmlRef<GdmlElCone> = register(GdmlElCone(name, dx, dy, zmax, zcut).apply(block))
+    ): GdmlRef<GdmlElCone> = register(GdmlElCone(resolveName<GdmlElCone>(name), dx, dy, zmax, zcut).apply(block))
 
     @GdmlApi
     public inline fun paraboloid(
-        name: String,
         rlo: Number,
         rhi: Number,
         dz: Number,
+        name: String? = null,
         block: GdmlParaboloid.() -> Unit = {},
-    ): GdmlRef<GdmlParaboloid> = register(GdmlParaboloid(name, rlo, rhi, dz).apply(block))
+    ): GdmlRef<GdmlParaboloid> = register(GdmlParaboloid(resolveName<GdmlParaboloid>(name), rlo, rhi, dz).apply(block))
 
     @GdmlApi
     public inline fun para(
-        name: String,
         x: Number,
         y: Number,
         z: Number,
+        name: String? = null,
         alpha: Number,
         theta: Number,
         phi: Number,
         block: GdmlParallelepiped.() -> Unit = {},
-    ): GdmlRef<GdmlParallelepiped> = register(GdmlParallelepiped(name, x, y, z, alpha, theta, phi).apply(block))
+    ): GdmlRef<GdmlParallelepiped> = register(GdmlParallelepiped(
+        resolveName<GdmlParallelepiped>(name),
+        x, y, z, alpha, theta, phi
+    ).apply(block))
 
     @GdmlApi
     public inline fun torus(
-        name: String,
         rmin: Number,
         rmax: Number,
         rtor: Number,
+        name: String? = null,
         block: GdmlTorus.() -> Unit = {},
-    ): GdmlRef<GdmlTorus> = register(GdmlTorus(name, rmin, rmax, rtor).apply(block))
+    ): GdmlRef<GdmlTorus> = register(GdmlTorus(resolveName<GdmlTorus>(name), rmin, rmax, rtor).apply(block))
 
     @GdmlApi
     public inline fun trd(
-        name: String,
         x1: Number,
         x2: Number,
         y1: Number,
         y2: Number,
         z: Number,
+        name: String? = null,
         block: GdmlTrapezoid.() -> Unit = {},
-    ): GdmlRef<GdmlTrapezoid> = register(GdmlTrapezoid(name, x1, x2, y1, y2, z).apply(block))
+    ): GdmlRef<GdmlTrapezoid> =
+        register(GdmlTrapezoid(resolveName<GdmlTrapezoid>(name), x1, x2, y1, y2, z).apply(block))
 
     @GdmlApi
     public inline fun polyhedra(
-        name: String,
         numsides: Int,
+        name: String? = null,
         block: GdmlPolyhedra.() -> Unit = {},
-    ): GdmlRef<GdmlPolyhedra> = register(GdmlPolyhedra(name, numsides).apply(block))
+    ): GdmlRef<GdmlPolyhedra> = register(GdmlPolyhedra(resolveName<GdmlPolyhedra>(name), numsides).apply(block))
 
     @GdmlApi
     public inline fun polycone(
-        name: String,
+        name: String? = null,
         block: GdmlPolycone.() -> Unit = {},
-    ): GdmlRef<GdmlPolycone> = register(GdmlPolycone(name).apply(block))
+    ): GdmlRef<GdmlPolycone> = register(GdmlPolycone(resolveName<GdmlPolycone>(name)).apply(block))
 
     @GdmlApi
     public inline fun scaledSolid(
-        name: String,
         solidref: GdmlRef<GdmlSolid>,
         scale: GdmlScale,
+        name: String? = null,
         block: GdmlScaledSolid.() -> Unit = {},
-    ): GdmlRef<GdmlScaledSolid> = register(GdmlScaledSolid(name, solidref, scale).apply(block))
+    ): GdmlRef<GdmlScaledSolid> =
+        register(GdmlScaledSolid(resolveName<GdmlScaledSolid>(name), solidref, scale).apply(block))
 
     @GdmlApi
-    public inline fun tube(name: String, rmax: Number, z: Number, block: GdmlTube.() -> Unit = {}): GdmlRef<GdmlTube> =
+    public inline fun tube(rmax: Number, z: Number, name: String, block: GdmlTube.() -> Unit = {}): GdmlRef<GdmlTube> =
         register(GdmlTube(name, rmax, z).apply(block))
 
     @GdmlApi
@@ -359,46 +378,46 @@ public class GdmlSolidContainer : GdmlContainer<GdmlSolid>() {
 
     @GdmlApi
     public inline fun cone(
-        name: String,
         z: Number,
         rmax1: Number,
         rmax2: Number,
+        name: String? = null,
         block: GdmlCone.() -> Unit = {},
     ): GdmlRef<GdmlCone> {
-        val cone = GdmlCone(name, z, rmax1, rmax2).apply(block)
+        val cone = GdmlCone(resolveName<GdmlCone>(name), z, rmax1, rmax2).apply(block)
         return register(cone)
     }
 
     @GdmlApi
     public inline fun union(
-        name: String,
         first: GdmlRef<GdmlSolid>,
         second: GdmlRef<GdmlSolid>,
+        name: String? = null,
         block: GdmlUnion.() -> Unit = {},
     ): GdmlRef<GdmlUnion> {
-        val union = GdmlUnion(name, first, second).apply(block)
+        val union = GdmlUnion(resolveName<GdmlUnion>(name), first, second).apply(block)
         return register(union)
     }
 
     @GdmlApi
     public inline fun intersection(
-        name: String,
         first: GdmlRef<GdmlSolid>,
         second: GdmlRef<GdmlSolid>,
+        name: String? = null,
         block: GdmlIntersection.() -> Unit = {},
     ): GdmlRef<GdmlIntersection> {
-        val intersection = GdmlIntersection(name, first, second).apply(block)
+        val intersection = GdmlIntersection(resolveName<GdmlIntersection>(name), first, second).apply(block)
         return register(intersection)
     }
 
     @GdmlApi
     public inline fun subtraction(
-        name: String,
         first: GdmlRef<GdmlSolid>,
         second: GdmlRef<GdmlSolid>,
+        name: String? = null,
         block: GdmlSubtraction.() -> Unit = {},
     ): GdmlRef<GdmlSubtraction> {
-        val subtraction = GdmlSubtraction(name, first, second).apply(block)
+        val subtraction = GdmlSubtraction(resolveName<GdmlSubtraction>(name), first, second).apply(block)
         return register(subtraction)
     }
 }
@@ -411,24 +430,22 @@ public class GdmlStructure : GdmlContainer<GdmlGroup>() {
 
     @GdmlApi
     public inline fun volume(
-        name: String,
         materialref: GdmlRef<GdmlMaterial>,
         solidref: GdmlRef<GdmlSolid>,
+        name: String? = null,
         block: GdmlVolume.() -> Unit = {},
-    ): GdmlVolume {
-        val res = GdmlVolume(name, materialref, solidref).apply(block)
-        add(res)
-        return res
+    ): GdmlRef<GdmlVolume> {
+        val res = GdmlVolume(resolveName<GdmlVolume>(name), materialref, solidref).apply(block)
+        return register(res)
     }
 
     @GdmlApi
     public inline fun assembly(
-        name: String,
+        name: String? = null,
         block: GdmlAssembly.() -> Unit = {},
-    ): GdmlAssembly {
-        val res = GdmlAssembly(name).apply(block)
-        add(res)
-        return res
+    ): GdmlRef<GdmlAssembly> {
+        val res = GdmlAssembly(resolveName<GdmlAssembly>(name)).apply(block)
+        return register(res)
     }
 }
 
@@ -447,6 +464,6 @@ public fun Gdml.world(
     block: GdmlAssembly.() -> Unit,
 ) {
     structure {
-        assembly(name, block).also { setup.world = it.ref() }
+        assembly(name, block).also { setup.world = it }
     }
 }